refactor(selection): remove DomBoundary class and integrate its functionality into SelectionController
- Deleted the `DomBoundary` class to simplify the selection handling logic.
- Integrated boundary-related methods directly into `SelectionController`, enhancing cohesion.
- Updated the README to reflect the removal of `DomBoundary` and clarify the new structure.
- Adjusted related imports and method calls throughout the selection feature to ensure consistency.
